% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlink2AssocFromPGEN.R
\name{calcPlink2AssocFromPGEN}
\alias{calcPlink2AssocFromPGEN}
\title{CALCULATE ASSOCIATION WITH PLINK 2.0 FROM PGEN FILE}
\usage{
calcPlink2AssocFromPGEN(
  sampleinfo_fn,
  covariables,
  adjust4sex = T,
  outcome,
  out_fn,
  keep_fn = "",
  snps2extract = "",
 
    pfiles = "/net/ifs1/san_projekte/projekte/genstat/02_projekte/1612_lifea1_genotypisierungsrunde_3/Imputation/03_imputed_prephased/bgen1_2/combined/s302_3_PLINK_QCed_StandardRS_combined_a1_rd3_chr1to23_N7660",
  unixcomputer = "amanMRO",
  use_firth_fallback = T,
  code_chrX = "m:0_2_f:0_1_2",
  showCovarStats = F,
  max_vif = 50,
  morePlinkParameter = " --threads 2 --memory 6000",
  gz = T,
 
    plinklocation_avx = "/net/ifs1/san_projekte/projekte/genstat/07_programme/plink2.0/20201105/unix_64/plink2",
 
    plinklocation_unix64 = "/net/ifs1/san_projekte/projekte/genstat/07_programme/plink2.0/20201105/unix_avx2/plink2"
)
}
\arguments{
\item{sampleinfo_fn}{File with outcome variable and covariates mit phenotypen, first two columns used as ID, must include the same IDs as defined in the psam file specified in \code{pfiles} parameter}

\item{covariables}{colnames of sampleinfo_fn specifying confounder to be included in model}

\item{adjust4sex}{Shall be adjusted for Sex? Do not include Sex as covariables, as this might result in bad behaviour in X-chromosomal analysis, Default: T}

\item{outcome}{colnames of sampleinfo_fn specifying the outcome variable, can be binary or continous}

\item{out_fn}{directory and prefix of result filenames}

\item{keep_fn}{Filename of a file consisting of two columns used as ID of individuals to keep. \code{''} means no filtering, Default: ''}

\item{snps2extract}{Filename of a file including SNP IDs as specified in the pvar file specified in parameter \code{pfiles} to keep. \code{''} means no filtering, Default: ''}

\item{pfiles}{full path to psam, pgen, and pvar fileset including the filename without file extension, Default: "/net/ifs1/san_projekte/projekte/genstat/02_projekte/1612_lifea1_genotypisierungsrunde_3/Imputation/03_imputed_prephased/bgen1_2/combined/s302_3_PLINK_QCed_StandardRS_combined_a1_rd3_chr1to23_N7660"}

\item{unixcomputer}{name of compute server, required to choose right plink version. Can be 'amanMRO' that will lead to use of version specified in  plinklocation_avx, otherwise, plink version specified in plinklocation_unix64 is used, Default: 'amanMRO'}

\item{use_firth_fallback}{to stabilize low maf calculation in logistic regression, see https://www.cog-genomics.org/plink/2.0/assoc, Default: T}

\item{code_chrX}{encoding can be either \verb{m:0_2_f:0_1_2} or \verb{m:0_1_f:0_1_2}, Default: 'm:0_2_f:0_1_2'}

\item{showCovarStats}{Include association summary statistic of covariates in output file, Default: F}

\item{max_vif}{variance inflation factor still ok in plink regression, reflecting correlation of covariates, can be calculated for a model using car::vif(mymodel) , Default: 50}

\item{morePlinkParameter}{plink parameters added to the call, see https://www.cog-genomics.org/plink/2.0/, Default: ' --threads 2 --memory 6000'}

\item{gz}{compress result files to gz, Default: T}

\item{plinklocation_avx}{Location of plink executable requiring unix 64 bit architecture with avx support, Default: '/net/ifs1/san_projekte/projekte/genstat/07_programme/plink2.0/20201105/unix_64/plink2'}

\item{plinklocation_unix64}{Location of plink executable requiring any unix 64 bit architecture, Default: '/net/ifs1/san_projekte/projekte/genstat/07_programme/plink2.0/20201105/unix_avx2/plink2'}
}
\value{
Returns a data frame with standard plink association results, see also https://www.cog-genomics.org/plink/2.0/formats#glm_linear or https://www.cog-genomics.org/plink/2.0/formats#glm_logistic Additionally all plink output is saved in the location specified in \code{out_fn}
}
\description{
Uses a psam file and plink 2.0 to quickly calculate associations for quantitative and binary outcome variable controlling for covariates for autosomes and the X-chromosome
}
\details{
Does not use rounded genotypes, but imputed gene doses. Works currently only on unix computeserver of GEnstat. Function in progress. Note that the default PGEN File (\verb{/net/ifs1/san_projekte/projekte/genstat/02_projekte/1612_lifea1_genotypisierungsrunde_3/Imputation/03_imputed_prephased/bgen1_2/combined/s302_3_PLINK_QCed_StandardRS_combined_a1_rd3_chr1to23_N7660}) includes only 7660 from LIFE-ADULT, excluding 9 Individuals that were QC ok for autosomes, but QC not ok for gonosomes
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
