% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansumstats.R
\name{clean_sumstats}
\alias{clean_sumstats}
\title{Title}
\usage{
clean_sumstats(infile, col_map, name, model = "logistic")
}
\arguments{
\item{infile}{raw gwas sumstat to clean}

\item{col_map}{map of column names}

\item{name}{gwas sumstat name}

\item{model}{logistic/linear}
}
\value{
a filepath to the slurm job
}
\description{
Title
}
\examples{
\dontrun{
clean_sumstats("my_raw_gwas.tsv.gz",name =  "mdd2024",col_map = col_map)
}
}
