% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma.R
\name{cleansumstats_magma}
\alias{cleansumstats_magma}
\title{Generate code to run magma gene assoc + gene_set assoc}
\usage{
cleansumstats_magma(
  paths,
  geneset_annotation_files = c(Sys.getenv("siletti_magma_clusters"),
    Sys.getenv("siletti_magma_superclusters"))
)
}
\arguments{
\item{paths}{list of filepaths, defined using filepath_manager()}

\item{geneset_annotation_files}{character vector of filepaths to gene annotations}
}
\value{
a slurm job (as a character vector)
}
\description{
Generate code to run magma gene assoc + gene_set assoc
}
\examples{
\dontrun{
cleansumstats_magma(filepath_manager("test_gwas"))
}

}
