% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansumstats.R
\name{correct_small_pval}
\alias{correct_small_pval}
\title{Represent very small pvalues as minimum double number}
\usage{
correct_small_pval(path)
}
\arguments{
\item{path}{filepath for gwas file}
}
\value{
writes out the gwas file
}
\description{
Represent very small pvalues as minimum double number
}
\examples{
\dontrun{
correct_small_pval("/my_gwas/gwas.tsv.gz")
}
}
