% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansumstats.R
\name{filepath_manager}
\alias{filepath_manager}
\title{Title}
\usage{
filepath_manager(infile, name, dir)
}
\arguments{
\item{infile}{gwas sumstat to clean}

\item{name}{optional: name to call the sumstat}

\item{dir}{directory used to store cleaned sumstats}
}
\value{
a list of filepaths
}
\description{
Title
}
\examples{
\dontrun{
path <- filepath_manager("adhd_2023.tsv.gz", "adhd_pgc2023")
}
}
