% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_cleaned_sumstats}
\alias{get_cleaned_sumstats}
\title{Get filepath and dataset_name for all cleaned sumstats}
\usage{
get_cleaned_sumstats(dir = Sys.getenv("gwasHelper_repo"))
}
\arguments{
\item{dir}{filepath for the directory to search. Defaults to use the ENV
variable "gwasHelper_repo"}
}
\value{
a tibble
}
\description{
Get filepath and dataset_name for all cleaned sumstats
}
\examples{
\dontrun{
# most common use is to call without arugemnt
get_cleaned_sumstats()
}
}
