% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma.R
\name{magma_munge_and_setup}
\alias{magma_munge_and_setup}
\title{Create the nessecary field for MAGMA gene & gene-set analysis}
\usage{
magma_munge_and_setup(input_gwas, output_dir)
}
\arguments{
\item{input_gwas}{filepath to input GWAS}

\item{output_dir}{output directory of results}
}
\value{
Returns nothing, but writes files to output dir when called
}
\description{
Create the nessecary field for MAGMA gene & gene-set analysis
}
\examples{
\dontrun{
magma_munge_and_setup("my_gwas", "/gwas_project/magma/version2")
}
}
