% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansumstats.R
\name{match_cols}
\alias{match_cols}
\title{Title}
\usage{
match_cols(
  col_CHR = NULL,
  col_POS = NULL,
  col_BETA = NULL,
  col_SNP = NULL,
  col_EffectAllele = NULL,
  col_SE = NULL,
  col_Z = NULL,
  col_OR = NULL,
  col_N = NULL,
  col_CaseN = NULL,
  col_ControlN = NULL,
  col_EAF = NULL,
  col_INFO = NULL,
  col_StudyN = NULL,
  col_OtherAllele = NULL,
  col_P = NULL,
  stats_CaseN = NULL,
  stats_ControlN = NULL,
  stats_StudyN = NULL
)
}
\arguments{
\item{col_CHR}{chromosome}

\item{col_POS}{position}

\item{col_BETA}{effect column (beta or odds ratio)}

\item{col_SNP}{rsid column}

\item{col_EffectAllele}{effect allele}

\item{col_SE}{standard error}

\item{col_Z}{Z score}

\item{col_OR}{odds ratio}

\item{col_N}{sample size}

\item{col_CaseN}{case sample size}

\item{col_ControlN}{control sample size}

\item{col_EAF}{allele frequency of effect allele}

\item{col_INFO}{info column}

\item{col_StudyN}{development column, dont use}

\item{col_OtherAllele}{other allele}

\item{col_P}{P value}

\item{stats_CaseN}{case sample size for entire study(not per snp)}

\item{stats_ControlN}{control sample size for entire study (not per snp)}

\item{stats_StudyN}{(sample size for entire study - not per snp)}
}
\value{
a col_map to use with clean_sumstats
}
\description{
Title
}
\examples{
\dontrun{
col_map(col_CHR = "CHR")
}

}
