% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_pldsc}
\alias{parse_pldsc}
\title{Read in the results of partitioned LDSC into tidy format}
\usage{
parse_pldsc(path)
}
\arguments{
\item{path}{to .results file from partitioned LDSC}
}
\value{
a tibble with one row: results for L2_L1 annotation
}
\description{
Read in the results of partitioned LDSC into tidy format
}
\examples{
\dontrun{
parse_pldsc("/pldsc/amygdala_excitatory.Results")
}
}
