% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayes.R
\name{prepare_ma}
\alias{prepare_ma}
\title{Title}
\usage{
prepare_ma(path, out, N, freq_path = Sys.getenv("freq1000g"))
}
\arguments{
\item{path}{filepath for sumstat}

\item{out}{filepath for .ma file}

\item{N}{Sample size - optional to use if missing N}

\item{freq_path}{path to SNP, A1, A2 freq if allele frequency is missing}
}
\value{
a character vector (commandline code)
}
\description{
Title
}
\examples{
\dontrun{
prepare_ma(paths)
}

}
