% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumping.R
\name{ranges_to_bed}
\alias{ranges_to_bed}
\title{Conver the clumping.ranges format from plink --clump to bed compatible format}
\usage{
ranges_to_bed(infile, out)
}
\arguments{
\item{infile}{a clumping.ranges file from plink --clump}

\item{out}{filename of output}
}
\value{
a character vector of commandline code
}
\description{
Conver the clumping.ranges format from plink --clump to bed compatible format
}
\examples{
\dontrun{
ranges_to_bed("analysis/plink/clumping/scz2022.clumping.ranges", "scz2022_clumping/ranges.bed")
}
}
