% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{run_pldsc}
\alias{run_pldsc}
\title{Generates commandline code to run partitioned ldscore regression}
\usage{
run_pldsc(sumstats, outdir, ldscores)
}
\arguments{
\item{sumstats}{ldsc munged sumstats to use}

\item{outdir}{output directory for results}

\item{ldscores}{a filepath to a directory with LD scores. Each folder in this directory
corresponds to the ldscores of the specific genes for that cell-type}
}
\value{
a character vector
}
\description{
Generates commandline code to run partitioned ldscore regression
}
\examples{
\dontrun{
run_pldsc("data/gwas/cognition.fastGWA.tsv", "results/partition_ldsc", "mouse_brain/LDSCORES")
}
}
