% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayes.R
\name{sbayes_run}
\alias{sbayes_run}
\title{Title}
\usage{
sbayes_run(paths)
}
\arguments{
\item{paths}{a list of filepaths, created with filepath_manager()}
}
\value{
a path to a .sh txt file
}
\description{
Title
}
\examples{
\dontrun{
sbayes_run(filepath_manager("adhd_2023.tsv.gz"))
}
}
