% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slurm_header}
\alias{slurm_header}
\title{Title}
\usage{
slurm_header(time, mem, output_dir, dependency = NULL)
}
\arguments{
\item{time}{--time parameter in slurm}

\item{mem}{--mem parameter in slurm}

\item{output_dir}{in what directory to put the slurm output file}

\item{dependency}{default null, if specified will add dependecy=afterok:{dependency}}
}
\value{
a character vector
}
\description{
Title
}
\examples{
slurm_header(24, 13, "my_dir/run_job")
}
