% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansumstats.R
\name{clean_sumstats}
\alias{clean_sumstats}
\title{Title}
\usage{
clean_sumstats(
  infile,
  col_map,
  name,
  model = "logistic",
  magma_gene = FALSE,
  magma_gene_set = FALSE,
  pldsc = FALSE,
  sbayesr = FALSE
)
}
\arguments{
\item{infile}{raw gwas sumstat to clean}

\item{col_map}{map of column names}

\item{name}{gwas sumstat name}

\item{model}{logistic/linear}

\item{magma_gene}{run magma gene analysis?}

\item{magma_gene_set}{run magma gene analysis?}

\item{pldsc}{run partitioned ldscore regression?}

\item{sbayesr}{run sbayesR to generate a PGS?}
}
\value{
a filepath to the slurm job
}
\description{
Title
}
\examples{
\dontrun{
clean_sumstats("my_raw_gwas.tsv.gz",name =  "mdd2024",col_map = col_map)
}
}
